requirements.txt
Flask==0.12.1
google-cloud-datastore==0.22.1
gunicorn==19.6.0
